moved shell to local utils module
