+++ src/app/commands/list.py
@@ -1,6 +1,6 @@
-from ..shell import *
 from .. import user_manager, login_disabled
 from ..obj import SystemMessenger, User, Command
+from utils.shell import Console
 
 SHL = Console("Command List")
