+++ src/app/sockets.py
@@ -8,9 +8,9 @@ from validators import url as val_url
 from . import socketio, emote_handler,  user_manager, verify_token, \
     emote_regex, html_regex, newline_html_regex, link_regex, youtube_regex, image_regex, video_regex, audio_regex, \
     code_regex, quote_regex, login_disabled, request, user_limit
-from .shell import *
 from . import handle_command as command_handler
 from .obj import User, Command, Message, get_default_user
+from utils.shell import Console, yellow2, white, green2
 
 SHL = Console("Socket")
